# docs/api/v1/swagger.yaml
basePath: /api/v1
definitions:
...
  discordgo.User:
    properties:
      accent_color:
        description: User's banner color, encoded as an integer representation of
          hexadecimal color code
        type: integer
      avatar:
        description: |-
          The hash of the user's avatar. Use Session.UserAvatar
          to retrieve the avatar itself.
        type: string
      banner:
        description: The hash of the user's banner image.
        type: string
      bot:
        description: Whether the user is a bot.
        type: boolean
      discriminator:
        description: The discriminator of the user (4 numbers after name).
        type: string
      email:
        description: |-
          The email of the user. This is only present when
          the application possesses the email scope for the user.
        type: string
      flags:
        description: |-
          The flags on a user's account.
          Only available when the request is authorized via a Bearer token.
        type: integer
      id:
        description: The ID of the user.
        type: string
      locale:
        description: The user's chosen language option.
        type: string
      mfa_enabled:
        description: Whether the user has multi-factor authentication enabled.
        type: boolean
      premium_type:
        description: |-
          The type of Nitro subscription on a user's account.
          Only available when the request is authorized via a Bearer token.
        type: integer
      public_flags:
        allOf:
        - $ref: '#/definitions/discordgo.UserFlags'
        description: |-
          The public flags on a user's account.
          This is a combination of bit masks; the presence of a certain flag can
          be checked by performing a bitwise AND between this int and the flag.
      system:
        description: Whether the user is an Official Discord System user (part of
          the urgent message system).
        type: boolean
      token:
        description: |-
          The token of the user. This is only present for
          the user represented by the current session.
        type: string
      username:
        description: The user's username.
        type: string
      verified:
        description: Whether the user's email is verified.
        type: boolean
    type: object
  discordgo.UserFlags:
    enum:
    - 1
    - 2
    - 4
    - 8
    - 64
    - 128
    - 256
    - 512
    - 1024
    - 4096
    - 16384
    - 65536
    - 131072
    - 262144
    type: integer
    x-enum-varnames:
    - UserFlagDiscordEmployee
    - UserFlagDiscordPartner
    - UserFlagHypeSquadEvents
    - UserFlagBugHunterLevel1
    - UserFlagHouseBravery
    - UserFlagHouseBrilliance
    - UserFlagHouseBalance
    - UserFlagEarlySupporter
    - UserFlagTeamUser
    - UserFlagSystem
    - UserFlagBugHunterLevel2
    - UserFlagVerifiedBot
    - UserFlagVerifiedBotDeveloper
    - UserFlagDiscordCertifiedModerator
...
  github_com_zekurio_kikuri_internal_services_webserver_v1_models.Error:
    properties:
      code:
        type: integer
      context:
        type: string
      error:
        type: string
    type: object
...
  github_com_zekurio_kikuri_internal_services_webserver_v1_models.Member:
    properties:
      avatar:
        description: The hash of the avatar for the guild member, if any.
        type: string
      avatar_url:
        type: string
      communication_disabled_until:
        description: |-
          The time at which the member's timeout will expire.
          Time in the past or nil if the user is not timed out.
        type: string
      created_at:
        type: string
      deaf:
        description: Whether the member is deafened at a guild level.
        type: boolean
      guild_id:
        description: The guild ID on which the member exists.
        type: string
      guild_name:
        type: string
      joined_at:
        description: The time at which the member joined the guild.
        type: string
      mute:
        description: Whether the member is muted at a guild level.
        type: boolean
      nick:
        description: The nickname of the member, if they have one.
        type: string
      pending:
        description: Is true while the member hasn't accepted the membership screen.
        type: boolean
      permissions:
        description: Total permissions of the member in the channel, including overrides,
          returned when in the interaction object.
        example: "0"
        type: string
      premium_since:
        description: When the user used their Nitro boost on the server
        type: string
      privilege:
        type: integer
      roles:
        description: A list of IDs of the roles which are possessed by the member.
        items:
          type: string
        type: array
      user:
        allOf:
        - $ref: '#/definitions/discordgo.User'
        description: The underlying user on which the member is based.
    type: object
  github_com_zekurio_kikuri_internal_services_webserver_v1_models.PermissionsResponse:
    properties:
      permissions:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  description: Kikuri API.
  title: Kikuri API
  version: "1.0"
paths:
...
  /guilds/{id}/{memberid}:
    get:
      consumes:
      - application/json
      description: Returns a single guild member by ID.
      parameters:
      - description: The ID of the guild.
        in: path
        name: id
        required: true
        type: string
      - description: The ID of the member.
        in: path
        name: memberid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_zekurio_kikuri_internal_services_webserver_v1_models.Member'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_zekurio_kikuri_internal_services_webserver_v1_models.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_zekurio_kikuri_internal_services_webserver_v1_models.Error'
      summary: Get Guild Member
      tags:
      - Members
  /guilds/{id}/{memberid}/permissions:
    get:
      consumes:
      - application/json
      description: Returns the permission array of the given user.
      parameters:
      - description: The ID of the guild.
        in: path
        name: id
        required: true
        type: string
      - description: The ID of the member.
        in: path
        name: memberid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_zekurio_kikuri_internal_services_webserver_v1_models.PermissionsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_zekurio_kikuri_internal_services_webserver_v1_models.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_zekurio_kikuri_internal_services_webserver_v1_models.Error'
      summary: Get Guild Member Permissions
      tags:
      - Members
  /guilds/{id}/{memberid}/permissions/allowed:
    get:
      consumes:
      - application/json
      description: Returns all detailed permission DNS which the member is alloed
        to perform.
      parameters:
      - description: The ID of the guild.
        in: path
        name: id
        required: true
        type: string
      - description: The ID of the member.
        in: path
        name: memberid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Wrapped in models.ListResponse
          schema:
            items:
              type: string
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_zekurio_kikuri_internal_services_webserver_v1_models.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_zekurio_kikuri_internal_services_webserver_v1_models.Error'
      summary: Get Guild Member Allowed Permissions
      tags:
      - Members
  /guilds/{id}/members:
    get:
      consumes:
      - application/json
      description: Returns a list of guild members.
      parameters:
      - description: The ID of the guild.
        in: path
        name: id
        required: true
        type: string
      - description: Request members after the given member ID.
        in: query
        name: after
        type: string
      - default: 100
        description: The amount of results returned.
        in: query
        maximum: 1000
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Wraped in models.ListResponse
          schema:
            items:
              $ref: '#/definitions/github_com_zekurio_kikuri_internal_services_webserver_v1_models.Member'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_zekurio_kikuri_internal_services_webserver_v1_models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_zekurio_kikuri_internal_services_webserver_v1_models.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_zekurio_kikuri_internal_services_webserver_v1_models.Error'
      summary: Get Guild Member List
      tags:
      - Members
...
swagger: "2.0"
...